@@ -12,10 +12,10 @@ classifiers =
     Programming Language :: Python :: 3
 
 [options]
-python_requires = >= 3.7
+python_requires = >= 3.11
 # https://setuptools.pypa.io/en/latest/userguide/dependency_management.html
 install_requires =
-    pyside6 ==6.2.4
+    pyside6
     qt-material
     QtAwesome
     cloudant
@@ -75,6 +75,7 @@ disallow_any_generics = true
 disallow_incomplete_defs = true
 disallow_untyped_defs = true
 warn_redundant_casts = true
+warn_unused_ignores = true
 ignore_missing_imports = true
 exclude = (?x)(
           pasta_eln/GUI/data_hierarchy/create_type_dialog_base.py
