@@ -2,7 +2,7 @@
 import platform, subprocess, os
 from enum import Enum
 from pathlib import Path
-from typing import Any
+from typing import Any, Union
 from PySide6.QtWidgets import QMenu, QWidget  # pylint: disable=no-name-in-module
 from PySide6.QtCore import     QPoint # pylint: disable=no-name-in-module
 from ..guiStyle import Action
@@ -18,13 +18,12 @@ def initContextMenu(widget:QWidget, pos:QPoint) -> None:
   """
   context = QMenu(widget)
   # for extractors
-  extractors = widget.comm.backend.configuration['extractors']
-  extension = Path(widget.doc['-branch'][0]['path']).suffix[1:]
+  extractors = widget.comm.backend.configuration['extractors']                                               # type: ignore[attr-defined]
+  extension = Path(widget.doc['-branch'][0]['path']).suffix[1:]                                              # type: ignore[attr-defined]
   if extension.lower() in extractors:
     extractors = extractors[extension.lower()]
-    baseDocType= widget.doc['-type'][0]
-    choices= {key:value for key,value in extractors.items() \
-                if key.startswith(baseDocType)}
+    baseDocType= widget.doc['-type'][0]                                                                      # type: ignore[attr-defined]
+    choices= {key:value for key,value in extractors.items() if key.startswith(baseDocType)}
     for key,value in choices.items():
       Action(value,                     widget, [CommandMenu.CHANGE_EXTRACTOR, key], context)
     context.addSeparator()
@@ -47,9 +46,9 @@ def executeContextMenu(widget:QWidget, command:list[Any]) -> bool:
   Returns:
     bool: success
   """
-  filePath = Path(widget.doc['-branch'][0]['path'])
+  filePath = Path(widget.doc['-branch'][0]['path'])                                                          # type: ignore[attr-defined]
   if command[0] is CommandMenu.OPEN_FILEBROWSER or command[0] is CommandMenu.OPEN_EXTERNAL:
-    filePath = widget.comm.backend.basePath/filePath
+    filePath = widget.comm.backend.basePath/filePath                                                         # type: ignore[attr-defined]
     filePath = filePath if command[0] is CommandMenu.OPEN_EXTERNAL else filePath.parent
     if platform.system() == 'Darwin':       # macOS
       subprocess.call(('open', filePath))
@@ -58,24 +57,25 @@ def executeContextMenu(widget:QWidget, command:list[Any]) -> bool:
     else:                                   # linux variants
       subprocess.call(('xdg-open', filePath))
   elif command[0] is CommandMenu.SAVE_IMAGE:
-    image = widget.doc['image']
+    image = widget.doc['image']                                                                              # type: ignore[attr-defined]
     if image.startswith('data:image/'):
       imageType = image[11:14] if image[14]==';' else image[11:15]
     else:
       imageType = 'svg'
-    saveFilePath = widget.comm.backend.basePath/filePath.parent/f'{filePath.stem}_PastaExport.{imageType.lower()}'
-    path = widget.doc['-branch'][0]['path']
+    saveFilePath = widget.comm.backend.basePath/filePath.parent/f'{filePath.stem}_PastaExport.{imageType.lower()}'# type: ignore[attr-defined]
+    path = widget.doc['-branch'][0]['path']                                                                  # type: ignore[attr-defined]
     if not path.startswith('http'):
-      path = (widget.comm.backend.basePath/path).as_posix()
-    widget.comm.backend.testExtractor(path, recipe='/'.join(widget.doc['-type']), saveFig=str(saveFilePath))
+      path = (widget.comm.backend.basePath/path).as_posix()                                                  # type: ignore[attr-defined]
+    widget.comm.backend.testExtractor(path, recipe='/'.join(widget.doc['-type']), saveFig=str(saveFilePath)) # type: ignore[attr-defined]
   elif command[0] is CommandMenu.HIDE:
-    widget.comm.backend.db.hideShow(widget.docID)
-    widget.doc = widget.comm.backend.db.getDoc(widget.docID)
+    widget.comm.backend.db.hideShow(widget.docID)                                                            # type: ignore[attr-defined]
+    widget.doc = widget.comm.backend.db.getDoc(widget.docID)                                                 # type: ignore[attr-defined]
   elif command[0] is CommandMenu.CHANGE_EXTRACTOR:
-    widget.doc['-type'] = command[1].split('/')
-    widget.comm.backend.useExtractors(filePath, widget.doc['shasum'], widget.doc)  #any path is good since the file is the same everywhere; data-changed by reference
-    if len(widget.doc['-type'])>1 and len(widget.doc['image'])>1:
-      widget.doc = widget.comm.backend.db.updateDoc({'image':widget.doc['image'], '-type':widget.doc['-type']}, widget.doc['_id'])
+    widget.doc['-type'] = command[1].split('/')                                                              # type: ignore[attr-defined]
+    #any path is good since the file is the same everywhere; data-changed by reference
+    widget.comm.backend.useExtractors(filePath, widget.doc['shasum'], widget.doc)                            # type: ignore[attr-defined]
+    if len(widget.doc['-type'])>1 and len(widget.doc['image'])>1:                                            # type: ignore[attr-defined]
+      widget.doc = widget.comm.backend.db.updateDoc({'image':widget.doc['image'], '-type':widget.doc['-type']}, widget.doc['_id'])# type: ignore[attr-defined]
     else:
       return False
   else:
